test(float, complex), chore(fuzz): PR #104 review follow-ups
Fixes from the review of the Mag/Ball + trig-family PR:

- `fuzz/run.sh` could report "all jobs passed" and exit 0 while every job
  failed: the log directory (`mktemp -d`) was unchecked and the failure
  sentinel lives inside it; a test binary whose `--ignored --list` failed was
  silently dropped; and `export FUZZ_CASES` shadowed the `PROPTEST_CASES`
  fallback the crate documents. All three are now checked/reported, `-p`
  entries are validated (a 0 width makes the complex tolerance vacuous), the
  `wait -n` pool throttle is probed once (bash 3.2 on macOS would otherwise
  launch every job at once), and `-k` prints the log path on failure too.
- `fuzz/src/lib.rs`: a zero case budget ran *no* case and passed — it now
  falls back to the default; the per-shard count is clamped rather than
  truncated by the `as u32`; and the precision sweep's subsampling is floored
  at `2^-2` (`MAX_SAMPLE_SHIFT`), so the widest width of the default sweep
  keeps a quarter of the cases instead of an eighth — an 8× sensitivity drop
  at exactly the width that exercises the multi-word kernels.
- `fuzz/tests/cmplx_transcendental.rs`: an `unnecessary_cast` clippy nit this
  branch introduced.
- `Ball::sqrt`: the documented `rad ≤ |mid|` precondition is now asserted in
  debug builds (the suite holds it at every width), and the zero-midpoint
  branch's contract is stated as it really is.
- `float/src/math/trig.rs`: the three targeted ×π tests swept a single
  precision each; they now cover the 20/50/100/500-bit set like the file's
  oracle tests.
- `float/src/mag.rs`: `exp_upper`'s justification said "minimum" where the
  inequality's extremum is a maximum (the bound is unaffected); the two
  comments naming another library's type/macro are rewritten in our own terms.
- `float/src/math/cache.rs`: `ln_base`'s generic-base range-error `expect` is
  documented in `# Panics` (unreachable for a base that fits a `Word`).
- `complex/src/math/trig.rs`: `# Examples` for the four new ×π convenience
  methods, matching what the same PR added to their `FBig` counterparts.

Tests, doc comments and a debug-only assertion: no changelog entry.

+++ fuzz/tests/cmplx_transcendental.rs
@@ -356,7 +356,7 @@ proptest! {
         for prec in fuzz::sampled_precisions_bits(fuzz::case_key(&[&zre, &zim, &n])) {
             let (z, rz) = pair(zre, zim, prec as usize);
             let d = cmplx_ok!(z.context().powi(&z, IBig::from(n)));
-            let r = rz.pow(&rug::Complex::with_val(prec as u32, (n as f64, 0.0)));
+            let r = rz.pow(&rug::Complex::with_val(prec, (n as f64, 0.0)));
             if !complex_finite(&d, &r) { continue; }
             prop_assert!(close_at(&d, &r, prec as usize), "powi zre={zre} zim={zim} n={n} prec={prec}");
         }
